+++ Test/verified_test.py
@@ -4,7 +4,7 @@
 >>> from math import isclose
 >>> from pycm import *
 >>> from pycm.overall_funcs import NIR_calc
->>> from pycm.pycm_ci import AUC_SE_calc, CI_calc
+>>> from pycm.ci import AUC_SE_calc, CI_calc
 >>> ABS_TOL = 1e-12
 >>> REL_TOL = 0
 >>> assert isclose(NIR_calc({'Class2': 804, 'Class1': 196}, 1000), 0.804, abs_tol=ABS_TOL, rel_tol=REL_TOL) # Verified Case - (Caret package)